add flexbox solution
Utilized flexbox to even the spacing between two columns in a row, as
well as dynamically center elements within one of the flex columns.
Padding/margins added to some of the elements to clean them up at
different screen sizes. Top/bottom margins still need to be added for
each row to aid in the spacing.